Implement Treseta game UI and functionality, including game setup, sorting controls, and history integration
