+++ server/src/main/resources/application-dev.properties
@@ -21,7 +21,9 @@ logging.level.org.hibernate.SQL=INFO
 logging.level.org.hibernate.orm.jdbc.bind=TRACE
 
 app.briskula-move.timer.duration-seconds=5
-app.treseta-move.timer.duration-seconds=5
+app.treseta-move.timer.duration-seconds=100
+app.lobby.timer.duration-seconds=300
+
 
 # Local cookies
 app.cookie-token.secure=false
