+++ server/src/main/resources/i18n/messages.properties
@@ -183,6 +183,9 @@ gameHistory.score=Score
 game.connectionLost=Connection lost. Reconnecting…
 game.waitingForGame=Waiting for game
 game.dragToPlay=Drag a card to play
+game.sort.aria=Sort cards
+game.sort.suit=Suit
+game.sort.rank=Rank
 game.waitingNextMove=Waiting for the next move
 game.card.alt=Card
 game.cardBack.alt=Card back
