+++ server/src/main/resources/i18n/messages_de.properties
@@ -177,6 +177,9 @@ gameHistory.score=Punktestand
 game.connectionLost=Verbindung verloren. Wiederverbindung…
 game.waitingForGame=Warten auf das Spiel
 game.dragToPlay=Ziehen Sie eine Karte, um zu spielen
+game.sort.aria=Karten sortieren
+game.sort.suit=Farbe
+game.sort.rank=Rang
 game.waitingNextMove=Warten auf den nächsten Zug
 game.card.alt=Karte
 game.cardBack.alt=Kartenrückseite
