+++ server/src/main/resources/i18n/messages_hr.properties
@@ -183,6 +183,9 @@ gameHistory.score=Rezultat
 game.connectionLost=Veza izgubljena. Ponovno spajanje…
 game.waitingForGame=Čekanje igre
 game.dragToPlay=Povucite kartu za igranje
+game.sort.aria=Sortiraj karte
+game.sort.suit=Vrsta
+game.sort.rank=Vrijednost
 game.waitingNextMove=Čekanje sljedećeg poteza
 game.card.alt=Karta
 game.cardBack.alt=Poleđina karte
