+++ server/src/main/resources/i18n/messages_uk.properties
@@ -183,6 +183,9 @@ gameHistory.score=Рахунок
 game.connectionLost=З'єднання втрачено. Повторне підключення…
 game.waitingForGame=Очікування гри
 game.dragToPlay=Перетягніть карту, щоб зіграти
+game.sort.aria=Сортувати карти
+game.sort.suit=Масть
+game.sort.rank=Ранг
 game.waitingNextMove=Очікування наступного ходу
 game.card.alt=Карта
 game.cardBack.alt=Сорочка карти
