+++ server/src/main/resources/templates/ui/fragments/game/live-game.html
@@ -11,6 +11,10 @@
                     <div id="player-summary-avatar" class="seat-avatar">P</div>
                     <div id="player-summary-name" class="player-summary-name"></div>
                     <div id="player-summary-points" class="player-summary-points">0</div>
+                    <div class="hand-sort-controls" th:aria-label="#{game.sort.aria}">
+                        <button id="hand-sort-suit" class="hand-sort-button" type="button" th:text="#{game.sort.suit}">Suit</button>
+                        <button id="hand-sort-rank" class="hand-sort-button" type="button" th:text="#{game.sort.rank}">Rank</button>
+                    </div>
                 </div>
             </div>
         </div>
